+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ### Changes
 - Neptune objects and universal methods covered with docstrings ([#1309](https://github.com/neptune-ai/neptune-client/pull/1309))
 
+### Features
+- Series objects accept `timestamps` and `steps` in their constructors ([#1318](https://github.com/neptune-ai/neptune-client/pull/1318))
+
 ## neptune 1.1.1
 
 ### Fixes